[MOV] im_livechat: move Feedback outside of public_livechat.js
Task-2818675
